Revert "Feat/ddp mlflow (#655)"
This reverts commit 5963d9ecd2acb72c6b11035e56d57f4fdcd8f022.
