+++ 3.test_cases/pytorch/cpu-ddp/Dockerfile
@@ -2,6 +2,6 @@ FROM pytorch/pytorch:latest
 
 RUN apt update && apt upgrade -y
 
-RUN pip install mlflow==2.13.2 sagemaker-mlflow==0.1.0
 COPY ddp.py /workspace
+
