+++ 3.test_cases/pytorch/cpu-ddp/kubernetes/fsdp.yaml-template
