+++ 3.test_cases/pytorch/cpu-ddp/slurm/2.create-enroot-image.sh
